Make simple unit test
