+++ package.json
@@ -30,6 +30,7 @@
   "devDependencies": {
     "@pxblue/eslint-config": "^1.1.5",
     "@pxblue/prettier-config": "^1.0.2",
+    "@types/jest": "^25.2.1",
     "@types/react-dom": "^16.9.7",
     "@typescript-eslint/eslint-plugin": "^2.31.0",
     "@typescript-eslint/parser": "^2.31.0",
